Create useIsDesktop hook
Create a custom hook for detecting if the screen is mobile size or desktop size to conditionally render navigation menus in other pages.
